error handling and HTTP responses for recipe ingredient controller

+++ PlatePlanner/src/main/java/org/launchcode/PlatePlanner/controller/RecipeIngredientController.java
@@ -3,7 +3,10 @@ package org.launchcode.PlatePlanner.controller;
 import jakarta.validation.Valid;
 import org.launchcode.PlatePlanner.model.RecipeIngredient;
 import org.launchcode.PlatePlanner.repository.RecipeIngredientRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.*;
 
@@ -15,35 +18,62 @@ import java.util.Optional;
 @CrossOrigin(origins = "http://localhost:5173")
 public class RecipeIngredientController {
 
+    Logger logger = LoggerFactory.getLogger(RecipeIngredientController.class);
+
     @Autowired
     private RecipeIngredientRepository recipeIngredientRepository;
 
     @GetMapping("/all")
-    public List<RecipeIngredient> getAllSavedRecipeIngredients() {
-        return (List<RecipeIngredient>) recipeIngredientRepository.findAll();
+    public ResponseEntity<List<RecipeIngredient>> getAllSavedRecipeIngredients() {
+        logger.info("In getAllSavedRecipeIngredients...");
+        return ResponseEntity.ok(recipeIngredientRepository.findAll());
     }
 
     @GetMapping("/{recipeIngredientId}")
-    public Optional<RecipeIngredient> getSavedRecipeIngredient(@PathVariable Long recipeIngredientId) {
-        return recipeIngredientRepository.findById(recipeIngredientId);
+    public ResponseEntity<Optional<RecipeIngredient>> getSavedRecipeIngredient(@PathVariable("recipeIngredientId") Long recipeIngredientId) {
+        logger.info("In getSavedRecipeIngredient...");
+        if (recipeIngredientRepository.existsById(recipeIngredientId)) {
+            logger.info("RecipeIngredient with ID {} found...", recipeIngredientId);
+            return ResponseEntity.ok(recipeIngredientRepository.findById(recipeIngredientId));
+        } else {
+            logger.warn("RecipeIngredient with ID {} not found...", recipeIngredientId);
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @PostMapping("/create")
-    public RecipeIngredient createRecipeIngredient(@RequestBody @Valid RecipeIngredient recipeIngredient, Errors errors) {
-        return recipeIngredientRepository.save(recipeIngredient);
+    public ResponseEntity<Object> createRecipeIngredient(@RequestBody @Valid RecipeIngredient recipeIngredient, Errors errors) {
+        logger.info("In createRecipeIngredient...");
+        recipeIngredientRepository.save(recipeIngredient);
+        if (errors.hasErrors()) {
+            logger.error("Error creating recipeIngredient: {}", errors);
+        }
+        return ResponseEntity.noContent().build();
     }
 
     @PostMapping("/update/{recipeIngredientId}")
-    public void updateRecipeIngredient(@PathVariable Long recipeIngredientId, @RequestBody @Valid RecipeIngredient recipeIngredient) {
+    public ResponseEntity<Object> updateRecipeIngredient(@PathVariable("recipeIngredientId") Long recipeIngredientId, @RequestBody @Valid RecipeIngredient recipeIngredient) {
+        logger.info("In updateRecipeIngredient...");
         if (recipeIngredientRepository.existsById(recipeIngredientId)) {
+            logger.info("RecipeIngredient with ID {} found.  Updating...", recipeIngredientId);
             recipeIngredientRepository.save(recipeIngredient);
+            return ResponseEntity.noContent().build();
         } else {
-//            TODO: throw error;
+            logger.warn("RecipeIngredient with ID {} not found...", recipeIngredientId);
+            return ResponseEntity.notFound().build();
         }
     }
 
     @DeleteMapping("/delete/{recipeIngredientId}")
-    public void deleteRecipeIngredient(@PathVariable Long recipeIngredientId) {
-        recipeIngredientRepository.deleteById(recipeIngredientId);
+    public ResponseEntity<Object> deleteRecipeIngredient(@PathVariable("recipeIngredientId") Long recipeIngredientId) {
+        logger.info("In deleteRecipeIngredient...");
+        if (recipeIngredientRepository.existsById(recipeIngredientId)) {
+            logger.info("RecipeIngredient with ID {} found.  Deleting...", recipeIngredientId);
+            recipeIngredientRepository.deleteById(recipeIngredientId);
+            return ResponseEntity.noContent().build();
+        } else {
+            logger.warn("RecipeIngredient with ID {} not found...", recipeIngredientId);
+            return ResponseEntity.notFound().build();
+        }
     }
 }
